fix(toss): 체결 수량 소수단위 절삭 방지 + 체결내역 정렬 보장
코드리뷰에서 나온 두 가지 확정 이슈를 고쳤다.

1. Fill.volume/FillItem.volume을 int에서 float으로 변경. 해외주식
   fractional 체결(예: 0.5주)이 int() 절삭으로 0주가 되어 대시보드
   체결내역에서 실제 체결이 사라지는 문제였다 — 옛 코드에도 있던 타입
   제약이지만, get_trades()가 항상 400으로 실패하던 지난 커밋 전에는
   실사용된 적이 없어 이번에 처음 노출됐다.
2. get_trades()가 "최신 체결 순"이라 문서화해놓고 실제로는 정렬을
   하지 않았다. Toss CLOSED 목록은 페이지 정렬 순서를 보장하지
   않으므로, 수집한 결과를 timestamp 내림차순으로 직접 정렬한다.

부수적으로 filled_at/ordered_at 파싱 중복을 _parse_toss_timestamp
헬퍼로 통합하고, 실패 시 조용히 넘어가지 않고 경고 로그를 남기도록
했다. 리뷰에서 지적된 테스트 갭(cursor 실제 전달 검증, count가 페이지
크기보다 작을 때 절단, CLOSED 결과 0건, 정렬 검증, fractional volume
보존)도 테스트로 추가했다.

+++ core/api/models.py
@@ -150,7 +150,9 @@ class FillItem(BaseModel):
 
     symbol: str
     price: Decimal
-    volume: int
+    # 해외주식 fractional investing 지원 — int면 0.5주 같은 체결이 0으로
+    # 잘린다 (core.adapters.toss.models.Fill.volume 참고).
+    volume: float
     timestamp: datetime
     currency: str
     side: Literal["BUY", "SELL"] | None = None
